Added support for Sims 4 Studio JSON files format
Sims 4 Studio output is technically not a loadable by the editor format, it's only used to represent the data in the editor. And exporting to the Sims 4 Studio JSON format is lacking other string table format information. This is only useful if other editors adopt the Sims 4 Studio JSON format which is the reason for this addition.